changing type to integer

+++ app/src/main/java/app/revanced/integrations/swipecontrols/SwipeControlsConfigurationProvider.kt
@@ -58,8 +58,8 @@ class SwipeControlsConfigurationProvider(
      * threshold for swipe detection
      * this may be called rapidly in onScroll, so we have to load it once and then leave it constant
      */
-    val swipeMagnitudeThreshold: Float
-        get() = SettingsEnum.SWIPE_MAGNITUDE_THRESHOLD.float
+    val swipeMagnitudeThreshold: Int
+        get() = SettingsEnum.SWIPE_MAGNITUDE_THRESHOLD.int
 //endregion
 
 //region overlay adjustments
@@ -79,8 +79,8 @@ class SwipeControlsConfigurationProvider(
     /**
      * text size for the overlay, in sp
      */
-    val overlayTextSize: Float
-        get() = SettingsEnum.SWIPE_OVERLAY_TEXT_SIZE.float
+    val overlayTextSize: Int
+        get() = SettingsEnum.SWIPE_OVERLAY_TEXT_SIZE.int
 
     /**
      * get the background color for text on the overlay, as a color int
